server/requirements.txt
websockets
pyautogui
pywebview
pystray
Pillow
requests